MCR-3355 improve class names for adapter implementations
